fix t1

+++ function_pointers/1-array_iterator.c
@@ -9,9 +9,8 @@
 */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int a;
+	unsigned int a;
 
 	for (a = 0; a < size; a++)
 	(*action)(array[a]);
 }
-
